+++ CHANGELOG.md
@@ -9,6 +9,16 @@
 ### Changed
 - Neo-to-Matrix conversion now supports converting Neo fields that have Super Table sub-field(s)
 
+## 2.8.14 - 2020-11-09
+
+> {warning} This release includes a migration affecting multi-site Craft installations, which reapplies propagation methods to blocks belonging to Neo fields with propagation methods other than "save blocks to all sites the owner element is saved in", due to a bug where changes to a field's propagation method were not being applied to their blocks.  If your Craft install is multi-site, make sure to backup your database before updating Neo.
+
+### Added
+- In Neo's JavaScript code, added the `afterInit` event to `Neo.Input`
+
+### Fixed
+- Fixed an issue on multi-site Craft installs where, when a Neo field's propagation method was changed, the field's blocks were not having the new propagation method applied
+
 ## 2.8.13 - 2020-11-02
 
 ### Changed